requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
alembic
python-dotenv
bcrypt==3.2.2
python-jose[cryptography]
pydantic[email]  # includes email-validator
requests
google-auth
passlib[bcrypt]